requirements.txt
fastapi
uvicorn

python-dotenv
pandas

asyncio
openai-agents

gradio
pydantic

agno
sqlalchemy

google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client